Improve package guard system to handle hotfixes
Most hotfixes requires a WMI query to detect whether they are installed
or not. Hotfixes maybe superseded by others and can't be installed when
a superseding package is already present.

To handle these case, update the semantic of the `not_if` field of the
packages definition:
* A String `not_if` value is a command to execute.
* A Array `not_if` value is a list of hotfixes to check using wmi.

Wmi-lite is used to perform a single query for all hotfixes. If any
of the defined hotfixes is present, the package won't be installed.
This should fix #45 & #54.

Tests and README.md have been updated accordingly.